+++ test/commands/setup_command_test.dart
@@ -110,6 +110,8 @@ void main() {
         'ios_capabilities',
         'branding',
         'screenshots',
+        'site',
+        'store',
       ]);
     });
   });